Moved zombie deck logic to it's own file
