+++ zombie_dice/zombie_dice.go
@@ -7,58 +7,15 @@ import (
 	"strconv"
 	"strings"
 	"time"
-
-	"github.com/Akavall/GoGamesProject/dice"
 )
 
-func initialize_deck() dice.Deck {
-
-	green := []string{"shot", "walk", "walk", "brain", "brain", "brain"}
-	yellow := []string{"shot", "shot", "walk", "walk", "brain", "brain"}
-	red := []string{"shot", "shot", "shot", "walk", "walk", "brain"}
-
-	green_sides := make_slice_of_sides(green)
-	yellow_sides := make_slice_of_sides(yellow)
-	red_sides := make_slice_of_sides(red)
-
-	// Put dices in the deck
-
-	const N_GREEN, N_YELLOW, N_RED = 6, 4, 3
-
-	dices := make([]dice.Dice, 0)
-
-	for i := 0; i < N_GREEN; i++ {
-		dices = append(dices, dice.Dice{Name: "green", Sides: green_sides})
-	}
-
-	for i := 0; i < N_YELLOW; i++ {
-		dices = append(dices, dice.Dice{Name: "yellow", Sides: yellow_sides})
-	}
-
-	for i := 0; i < N_RED; i++ {
-		dices = append(dices, dice.Dice{Name: "red", Sides: red_sides})
-	}
-
-	zombie_dice_deck := dice.Deck{Name: "ZombieDiceDeck", Dices: dices}
-
-	return zombie_dice_deck
-}
-
-func make_slice_of_sides(string_sides []string) []dice.Side {
-	sides := make([]dice.Side, len(string_sides))
-	for ind, s := range string_sides {
-		sides[ind] = dice.Side{Name: s}
-	}
-	return sides
-}
-
-func players_turn(deck dice.Deck, player_name string) (int, error) {
+func players_turn(deck ZombieDeck, player_name string) (int, error) {
 
 	brains := 0
 	shots := 0
 
 	for {
-		if len(deck.Dices) < 3 {
+		if len(deck.Deck.Dices) < 3 {
 			fmt.Printf("%s have ran out of dices", player_name)
 			fmt.Printf("%s final score is : %d", player_name, brains)
 			return brains, nil
@@ -128,7 +85,7 @@ func players_turn(deck dice.Deck, player_name string) (int, error) {
 func PlayWithAI() {
 	player_total_score := 0
 	ai_total_score := 0
-	deck := initialize_deck()
+	deck := InitZombieDeck()
 
 	ai_name := select_ai()
 
@@ -198,4 +155,3 @@ back:
 		goto back
 	}
 }
-
